+++ lib/datafoodconsortium/connector_v1/planned_local_transformation.rb
@@ -70,8 +70,8 @@ class DataFoodConsortium::ConnectorV1::PlannedLocalTransformation
 		registerSemanticProperty("dfc-b:cost", &method("cost")).valueSetter = method("cost=")
 		registerSemanticProperty("dfc-b:startDate", &method("startDate")).valueSetter = method("startDate=")
 		registerSemanticProperty("dfc-b:endDate", &method("endDate")).valueSetter = method("endDate=")
-		registerSemanticProperty("dfc-b:hasIncome", &method("consumptionFlow")).valueSetter = method("consumptionFlow=")
-		registerSemanticProperty("dfc-b:hasOutcome", &method("productionFlow")).valueSetter = method("productionFlow=")
+		registerSemanticProperty("dfc-b:hasInput", &method("consumptionFlow")).valueSetter = method("consumptionFlow=")
+		registerSemanticProperty("dfc-b:hasOutput", &method("productionFlow")).valueSetter = method("productionFlow=")
 	end