+++ spec/pickup_option_spec.rb
@@ -1,17 +1,17 @@
-RSpec.describe DataFoodConsortium::ConnectorV1::CustomerCategory do
+RSpec.describe DataFoodConsortium::ConnectorV1::PickupOption do
   it "can be empty when exported" do
-    subject = DataFoodConsortium::ConnectorV1::CustomerCategory.new(
+    subject = DataFoodConsortium::ConnectorV1::PickupOption.new(
       "https://myplatform.com/cc"
     )
     result = exported_json(subject)
     expect(result).to include(
       "@id" => "https://myplatform.com/cc",
-      "@type" => "dfc-b:CustomerCategory",
+      "@type" => "dfc-b:PickupOption",
     )
   end
 
   it "contains all fields when exported" do
-    subject = DataFoodConsortium::ConnectorV1::CustomerCategory.new(
+    subject = DataFoodConsortium::ConnectorV1::PickupOption.new(
       "https://myplatform.com/cc",
       name: "name",
       description: "description",
@@ -19,7 +19,7 @@ RSpec.describe DataFoodConsortium::ConnectorV1::CustomerCategory do
     result = exported_json(subject)
     expect(result).to include(
       "@id" => "https://myplatform.com/cc",
-      "@type" => "dfc-b:CustomerCategory",
+      "@type" => "dfc-b:PickupOption",
       "dfc-b:name" => "name",
       "dfc-b:description" => "description",
     )